app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
httpx
#DB
sqlalchemy
sqlalchemy_utils
psycopg2-binary
asyncpg
aiosqlite
python-jose
jwt
#mongoengine
#pymongo
#redis 
#broker
#rabbitmq
# migration 
#alembic
#liquibase
pytest
pytest-asyncio
python-dotenv
# alembic==1.14.1
# annotated-types==0.7.0
# anyio==4.8.0
# cffi==1.17.1
# click==8.1.8
# colorama==0.4.6
# dnspython==2.7.0
# fastapi==0.115.8
# greenlet==3.1.1
# h11==0.14.0
# idna==3.10
# Mako==1.3.8
# MarkupSafe==3.0.2
# mongoengine==0.29.1
# psycopg2-binary==2.9.10
# pycparser==2.22
# pydantic==2.10.6
# pydantic_core==2.27.2
# pymongo==4.11
# python-multipart==0.0.20
# rabbitmq==0.2.0
# redis==5.2.1
# sniffio==1.3.1
# SQLAlchemy==2.0.37
# starlette==0.45.3
# typing_extensions==4.12.2
# uvicorn==0.34.0